+++ app/Interfaces/DatabaseConnectionInterface.php
@@ -1,0 +1,7 @@
+<?php
+namespace App\Interfaces;
+
+interface DatabaseConnectionInterface
+{
+    public function getConnection();
+}