+++ index.php
@@ -1,15 +1,31 @@
 <?php
 require __DIR__ . '/vendor/autoload.php';
 
-define('ROOT', __DIR__);
+use Dotenv\Dotenv;
+
+// Load .env file from the root directory
+$dotenv = Dotenv::createImmutable(__DIR__);
+$dotenv->load();
+
+
+
 use App\Controllers\NewsController;
 use App\Controllers\CommentController;
 use App\Repositories\NewsRepository;
 use App\Repositories\CommentRepository;
+use App\Services\Connections\DatabaseConnectionService;
+use App\Services\DatabaseService;
 
-$commentRepository = new CommentRepository();
+$database_config = require_once "./config/database.php";
 
-foreach (NewsController::getInstance(new NewsRepository(), $commentRepository)->listNews() as $news) {
+
+$databaseConnection = new DatabaseConnectionService($database_config['mysql']);
+$databaseService = new DatabaseService($databaseConnection);
+
+$commentRepository = new CommentRepository($databaseService);
+$newsRepository = new NewsRepository($databaseService);
+
+foreach (NewsController::getInstance($newsRepository, $commentRepository)->listNews() as $news) {
 	echo ("############ NEWS " . $news->getTitle() . " ############\n");
 	echo ($news->getBody() . "\n");
 	foreach (CommentController::getInstance($commentRepository)->listComments() as $comment) {